change stock quantities to integer

+++ backend/app/models/item.py
@@ -14,5 +14,5 @@ class Item(Base):
     unit = Column(String(20), nullable=False)
     tax_rate = Column(Numeric(5, 2), nullable=False)
     opening_stock = Column(Numeric(10, 3), default=0)
-    current_stock = Column(Numeric(10, 3), default=0)
+    current_stock = Column(Integer, default=0)
     created_at = Column(DateTime, server_default=func.now())
